miniapp/zakaz/webapp/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
python-telegram-bot==20.4
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
cryptography==41.0.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
Pillow==10.0.0
python-dateutil==2.8.2 